video e script

+++ backEnd/model/entities/candidatosModel.js
@@ -5,11 +5,12 @@ const database = new DataBase();
 class CandidatosModel {
     async adicionarCandidato(candidato, connection) {
         try {
+            console.log('Model: Tentando adicionar candidato:', candidato);
             const [candidatoExiste] = await connection.query(`SELECT * FROM Candidato WHERE cand_cpf = ?`, [candidato.cpf]);
 
             if (candidatoExiste.length > 0) {
                 await database.rollbackTransaction(connection);
-                return { success: false, message: 'Candidato já cadastrado', candCpf: candidato.cpf };
+                return { success: false, message: 'Candidato já cadastrado', candCpf: candidato.cand_cpf };
             }
 
             const insertCandidatoQuery = `
@@ -17,14 +18,14 @@ class CandidatosModel {
                 VALUES (?, ?, ?, ?);
             `;
             await connection.query(insertCandidatoQuery, [
-                candidato.cpf,
-                candidato.nome,
-                candidato.endereco,
-                candidato.telefone
+                candidato.cand_cpf,
+                candidato.cand_nome,
+                candidato.cand_endereco,
+                candidato.cand_telefone
             ]);
 
             await database.commitTransaction(connection);
-            return { success: true, message: 'Candidato inserido com sucesso', candCpf: candidato.cpf };
+            return { success: true, message: 'Candidato inserido com sucesso', candCpf: candidato.cand_cpf };
         } catch (error) {
             await database.rollbackTransaction(connection);
             throw error;
@@ -61,14 +62,24 @@ class CandidatosModel {
     async excluirCandidato(cpf) {
         const connection = await database.beginTransaction();
         try {
-            const deleteCandidatoQuery = `DELETE FROM Candidato WHERE cand_cpf = ?`;
-            const [result] = await connection.query(deleteCandidatoQuery, [cpf]);
+            const candidatoExiste = await database.executaComando(`SELECT * FROM Candidato WHERE cand_cpf = ?`, [cpf]);
+            if (candidatoExiste.length === 0) {
+                await database.rollbackTransaction(connection);
+                return { success: false, message: 'Candidato não encontrado' };
+            }
+            const deleteCandidatoVagaQuery = `DELETE FROM Candidato_Vaga WHERE cand_cpf = ?`;
+            const [resultVaga] = await connection.query(deleteCandidatoVagaQuery, [cpf]);
 
-            await database.commitTransaction(connection);
-            if (result.affectedRows > 0) {
+            const deleteCandidatoQuery = `DELETE FROM Candidato WHERE cand_cpf = ?`;
+            const [resultCandidato] = await connection.query(deleteCandidatoQuery, [cpf]);
+
+            
+            if (resultCandidato.affectedRows > 0) {
+                await database.commitTransaction(connection);
                 return { success: true, message: 'Candidato excluído com sucesso' };
             } else {
-                return { success: false, message: 'Candidato não encontrado' };
+                await database.rollbackTransaction(connection);
+                return { success: false, message: 'Erro ao excluir candidato' };
             }
         } catch (error) {
             await database.rollbackTransaction(connection);
@@ -91,7 +102,12 @@ class CandidatosModel {
                 INSERT INTO Candidato_Vaga (cand_cpf, vaga_codigo, data_inscricao, horario_inscricao)
                 VALUES (?, ?, CURDATE(), CURTIME());
             `;
-            await connection.query(insertCandidatoVagaQuery, [candCpf, vagaCodigo]);
+            const [result] = await connection.query(insertCandidatoVagaQuery, [candCpf, vagaCodigo]);
+
+            if (result.affectedRows === 0) {
+                await database.rollbackTransaction(connection);
+                return { success: false, message: 'Candidatura não realizada' };
+            }
 
             await database.commitTransaction(connection);
             return { success: true, message: `Candidatura realizada com sucesso para a vaga ${vagaCodigo}` };
